+++ repository/entity/notification.go
@@ -1,4 +1,4 @@
-package model
+package entity
 
 import (
 	"time"
